+++ service/generic_swap_server.js
@@ -4,13 +4,17 @@ const express = require('express');
 const defaultSwapInBaseFee = 5000;
 const defaultSwapInFeeRate = 1500;
 const defaultSwapInCltv = 1000;
+const defaultSwapOutBaseFee = 500;
+const defaultSwapOutFeeRate = 500;
 const defaultMaxSwapInTokens = 30e6;
 const defaultMinSwapInTokens = 250e3;
 const defaultMaxSwapOutCltv = 400;
 const defaultMaxSwapOutTokens = 30e6;
 const defaultMinSwapOutCltv = 100;
 const defaultMinSwapOutTokens = 250e3;
+const defaultSwapOutDeposit = 30e3;
 const feeRateDenominator = BigInt(1e6);
+const methodNewSwapIn = 'newLoopInSwap';
 const methodSwapInQuote = 'loopInQuote';
 const methodSwapInTerms = 'loopInTerms';
 const methodSwapOutPushPreimage = 'loopOutPushPreimage';
@@ -21,6 +25,8 @@ const path = method => `/v0/${method}`;
 /** Create generic swap server
 
   {
+    destination: <Destination Hex String>
+    handle_swap_in: <Respond to Swap In Promise>
     [max_swap_in_tokens]: <Maximum Swap In Tokens Number>
     [min_swap_in_tokens]: <Minimum Swap In Tokens Number>
     [max_swap_out_cltv]: <Maximum Swap Out CLTV Delta Number>
@@ -30,6 +36,9 @@ const path = method => `/v0/${method}`;
     [swap_in_base_fee]: <Swap In Base Fee Tokens Number>
     [swap_in_fee_rate]: <Swap In Fee Rate Parts Per Million Number>
     [swap_in_cltv]: <Swap In CLTV Delta Number>
+    [swap_out_base_fee]: <Swap Out Base Fee Tokens Number>
+    [swap_out_deposit]: <Swap Out Deposit Tokens Number>
+    [swap_out_fee_rate]: <Swap Out Fee Rate Parts Per Million Number>
   }
 
   @returns
@@ -42,6 +51,14 @@ module.exports = args => {
 
   app.use(bodyParser.json());
 
+  app.post(path(methodNewSwapIn), async (req, res) => {
+    try {
+      return res.json(await args.handle_swap_in({}));
+    } catch (err) {
+      return res.json(err);
+    }
+  });
+
   app.post(path(methodSwapInQuote), (req, res) => {
     const swapInBaseFee = args.swap_in_base_fee || defaultSwapInBaseFee;
     const swapInFeeRate = args.swap_in_fee_rate || defaultSwapInFeeRate;
@@ -51,7 +68,7 @@ module.exports = args => {
 
     return res.json({
       cltv_delta: args.swap_in_cltv || defaultSwapInCltv,
-      swap_fee: swapInBaseFee + Number(rate),
+      swap_fee: (swapInBaseFee + Number(rate)).toString(),
     });
   });
 
@@ -63,6 +80,20 @@ module.exports = args => {
   });
 
   app.post(path(methodSwapOutPushPreimage), (req, res) => res.json({}));
+
+  app.post(path(methodSwapOutQuote), (req, res) => {
+    const swapOutBaseFee = args.swap_out_base_fee || defaultSwapOutBaseFee;
+    const swapOutFeeRate = args.swap_out_fee_rate || defaultSwapOutFeeRate;
+    const tokens = Number(req.body.amt);
+
+    const rate = BigInt(tokens) * BigInt(swapOutFeeRate) / feeRateDenominator;
+
+    return res.json({
+      prepay_amt: args.swap_out_deposit || defaultSwapOutDeposit,
+      swap_fee: (swapOutBaseFee + Number(rate)).toString(),
+      swap_payment_dest: args.destination,
+    });
+  });
 
   app.post(path(methodSwapOutTerms), (req, res) => {
     return res.json({