+++ test/service/test_get_swap_out_terms.js
@@ -5,7 +5,7 @@ const {genericSwapServer} = require('./../../service');
 const {genericSwapService} = require('./../../');
 const {getSwapOutTerms} = require('./../../');
 
-const port = 2349;
+const port = 2352;
 
 const socket = `http://localhost:${port}`;